Rename the api and ssh modules
